Create HW3.java
alphabeticalOrder–checks if a string is in alphabetical order
isWindingPath–given an double array of integers, it should be able to tell if it forms a winding path. For it to be a winding path, you must be able to start at the smallest number, go the the next largest number next to the smallest number, and repeat until you have reached all the values in the array. If you cannot reach all the values in the array, it is not a winding path.
10 11 14 15
 9 12 13
 8  1   2  3
 7  6  5  4

repeatChars–repeats the characters in a string by a certain number, returns a string with the repeated characters [i.e repeatChars("Ashley Sah", 3) returns "AAAssshhhllleeeyyy SSSaaahhh")

repeatWords–repeats the words in a string by a certain number (i.e repeatWords("Ashley Sah",2 ) returns "Ashley Ashley Sah Sah")

caesarShift()–shifts all the values in a string by an integer value, if the shift value a letter past 'Z' or a number past '9', it will loop back around 
cryptoQuipp()–picks a random shift value for each character in a string (so one character might be shifted by 2, while the next one may be shifted by 3)